+++ packages/www/src/repl/codemirror.js
@@ -1,0 +1,11 @@
+import codemirror from "codemirror";
+import "codemirror/mode/css/css.js";
+
+const theme = "base16-light";
+
+export default codemirror;
+
+export {
+    codemirror,
+    theme,
+};