+++ packages/www/src/repl/index.js
@@ -1,0 +1,5 @@
+import Repl from "./repl.html";
+
+new Repl({
+    target : document.body
+});